+++ app/includer/routers.py
@@ -1,5 +1,6 @@
 from entities.category.routers_category import router_category
 from entities.offer.routers_offer import router_offer
+from entities.order.routers_order import router_order
 from entities.service_name.routers_service_name import router_service
 from entities.users.routers.client_routers import router_client
 from entities.users.routers.employee_routers import router_employee
@@ -10,4 +11,5 @@ routers_all2 = [
     router_employee,
     router_service,
     router_offer,
+    router_order,
 ]
