+++ app/routers/__init__.py
@@ -1,11 +1,9 @@
 from config import i_config
 from routers.index import router_index
-from routers.order_routers import router_order
 from routers.stub_init_routers import router_init_stub
 
 routers_all = [
     router_index,
-    router_order,
 ]
 if i_config.DEBUG:
     routers_all.append(router_init_stub)
